fix timestamps and edit mapping

+++ MessageService/T_03/src/main/java/suraifokkusu/message/repositories/MessageRepository.java
@@ -1,12 +1,9 @@
 package suraifokkusu.message.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import suraifokkusu.message.entities.MessageEntity;
 
-import java.util.Optional;
-
 @Repository
-public interface MessageRepository extends JpaRepository<MessageEntity, Long> {
+public interface MessageRepository extends JpaRepository<MessageEntity, Integer> {
 }
